refactor(flow): move getJobStatus + listJobs read logic into PostgresJobQueue adapter (BR26 Lot 7 slice A)
